feat: react router prompt
